requirements.txt
fastapi 
uvicorn 
python-dotenv
gunicorn 
google-auth 
google-auth-oauthlib 
google-api-python-client
selenium
webdriver-manager
jsonify
joblib
celery
pillow
flask
sqlalchemy
pandas
google-analytics-data
httpx
python-multipart
requests
sendgrid